+++ Jinaga.Store.SQLite/Description/ExistentialConditionDescription.cs
@@ -1,0 +1,7 @@
+﻿namespace Jinaga.Store.SQLite.Description
+{
+    internal class ExistentialConditionDescription
+    {
+
+    }
+}
