+++ Jinaga.Store.SQLite/Description/InputDescription.cs
@@ -1,0 +1,20 @@
+﻿namespace Jinaga.Store.SQLite.Description
+{
+    internal class InputDescription
+    {
+        public string Label { get; }
+        public string Type { get; }
+        public int FactIndex { get; }
+        public int FactTypeParameter { get; }
+        public int FactHashParameter { get; }
+
+        public InputDescription(string label, string type, int factIndex, int factTypeParameter, int factHashParameter)
+        {
+            Label = label;
+            Type = type;
+            FactIndex = factIndex;
+            FactTypeParameter = factTypeParameter;
+            FactHashParameter = factHashParameter;
+        }
+    }
+}
