+++ Jinaga.Store.SQLite/Description/ResultDescription.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Immutable;
+
+namespace Jinaga.Store.SQLite.Description
+{
+    internal class ResultDescription
+    {
+        public QueryDescription QueryDescription { get; set; }
+        public ImmutableDictionary<string, ResultDescription> ChildResultDescriptions { get; set; }
+    }
+}
